requirements.txt
fastapi
uvicorn
langgraph
langchain
langchain-community
langchain-openai
langchain-groq
groq
python-dotenv
duckduckgo-search
tavily-python
pydantic
streamlit
pytest
aiohttp
requests
